Match members contract output to billing contract input
