+++ demo/data/dynamic_tariff.yaml
@@ -1,13 +1,23 @@
 # see https://github.com/EnergieID/energy-cost for more details on the tariff configuration format
-- start: 2025-01-01T00:00:00+01:00
+- start: 2024-01-01T00:00:00+01:00
   consumption:
     constant_cost: 10.0
     variable_costs:
     - index: Belpex15min
       scalar: 1.05
+  injection:
+    constant_cost: 14.56
+    variable_costs:
+    - index: Belpex15min
+      scalar: -0.9876
 - start: 2026-01-01T00:00:00+01:00
   consumption:
     constant_cost: 12.0
     variable_costs:
     - index: Belpex15min
       scalar: 1.10
+  injection:
+    constant_cost: 14.56
+    variable_costs:
+    - index: Belpex15min
+      scalar: -0.9876
