+++ demo/data/fixed_tariff.yaml
@@ -1,0 +1,25 @@
+# Fixed tariff configuration
+# See https://github.com/EnergieID/energy-cost for the format reference
+- start: 2024-01-01T00:00:00+01:00
+  consumption:
+    constant_cost: 100.0
+  injection:
+    constant_cost: -20.0
+- start: 2025-01-01T00:00:00+01:00
+  consumption:
+    constant_cost: 90.0
+  injection:
+    constant_cost: -15.0
+  periodic:
+    fixed_fee:
+      period: monthly
+      constant_cost: 10
+- start: 2026-01-01T00:00:00+01:00
+  consumption:
+    constant_cost: 120.0
+  injection:
+    constant_cost: -25.0
+  periodic:
+    fixed_fee:
+      period: monthly
+      constant_cost: 15
