+++ demo/data/monthly_index.csv
@@ -1,0 +1,13 @@
+timestamp,value
+2024-01-01T00:00:00+01:00,50.0
+2024-02-01T00:00:00+01:00,55.0
+2024-03-01T00:00:00+01:00,53.0
+2024-04-01T00:00:00+01:00,60.0
+2024-05-01T00:00:00+01:00,58.0
+2024-06-01T00:00:00+01:00,62.0
+2024-07-01T00:00:00+01:00,65.0
+2024-08-01T00:00:00+01:00,63.0
+2024-09-01T00:00:00+01:00,61.0
+2024-10-01T00:00:00+01:00,59.0
+2024-11-01T00:00:00+01:00,57.0
+2024-12-01T00:00:00+01:00,56.0
