+++ src/cofy/enums.py
@@ -1,0 +1,13 @@
+from enum import StrEnum
+
+
+class CustomerType(StrEnum):
+    RESIDENTIAL = "residential"
+    NON_RESIDENTIAL = "non_residential"
+    PROTECTED = "protected"
+
+
+class ConnectionType(StrEnum):
+    ELECTRICITY = "electricity"
+    GAS = "gas"
+    WATER = "water"
